set up pseudo-transaction docs, circular import issue
