+++ docs/source/xrpl.models.transactions.pseudo_transactions.rst
@@ -1,29 +1,8 @@
-xrpl.models.transactions.pseudo\_transactions package
-=====================================================
-
-Submodules
-----------
-
-xrpl.models.transactions.pseudo\_transactions.pseudo\_transaction module
-------------------------------------------------------------------------
-
-.. automodule:: xrpl.models.transactions.pseudo_transactions.pseudo_transaction
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.transactions.pseudo\_transactions.pseudo\_transaction\_type module
-------------------------------------------------------------------------------
-
-.. automodule:: xrpl.models.transactions.pseudo_transactions.pseudo_transaction_type
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
+XRPL Pseudo-Transaction Models
+==============================
 
 .. automodule:: xrpl.models.transactions.pseudo_transactions
    :members:
    :undoc-members:
    :show-inheritance:
+   :inherited-members:
