+++ xrpl/models/transactions/offer_cancel.py
@@ -2,7 +2,8 @@
 from dataclasses import dataclass, field
 
 from xrpl.models.required import REQUIRED
-from xrpl.models.transactions.transaction import Transaction, TransactionType
+from xrpl.models.transactions.transaction import Transaction
+from xrpl.models.transactions.types import TransactionType
 from xrpl.models.utils import require_kwargs_on_init
