+++ xrpl/models/transactions/payment_channel_claim.py
@@ -3,7 +3,8 @@ from dataclasses import dataclass, field
 from typing import Optional
 
 from xrpl.models.required import REQUIRED
-from xrpl.models.transactions.transaction import Transaction, TransactionType
+from xrpl.models.transactions.transaction import Transaction
+from xrpl.models.transactions.types import TransactionType
 from xrpl.models.utils import require_kwargs_on_init
