+++ xrpl/models/transactions/pseudo_transactions/__init__.py
@@ -2,3 +2,9 @@
 Model objects for specific `types of Pseudo-Transactions
 <https://xrpl.org/pseudo-transaction-types.html>`_ in the XRP Ledger.
 """
+from xrpl.models.transactions.pseudo_transactions.enable_amendment import (
+    EnableAmendment,
+    EnableAmendmentFlag,
+)
+
+__all__ = ["EnableAmendment", "EnableAmendmentFlag"]
