+++ xrpl/models/transactions/set_regular_key.py
@@ -2,7 +2,8 @@
 from dataclasses import dataclass, field
 from typing import Optional
 
-from xrpl.models.transactions.transaction import Transaction, TransactionType
+from xrpl.models.transactions.transaction import Transaction
+from xrpl.models.transactions.types import TransactionType
 from xrpl.models.utils import require_kwargs_on_init
